pyproject.toml
# ==== pytest ====
[tool.pytest.ini_options]
minversion = "6.0"
addopts = "--ds=config.settings.test --reuse-db --import-mode=importlib"
python_files = [
    "tests.py",
    "test_*.py",
]
filterwarnings = [
    "ignore::UserWarning:django.core.handlers.base",
    "ignore::pydantic.warnings.PydanticDeprecatedSince20:ninja",
]

# ==== Coverage ====
[tool.coverage.run]
include = ["temp_competibee/**"]
omit = ["*/migrations/*", "*/tests/*"]
plugins = ["django_coverage_plugin"]

# ==== mypy ====
[tool.mypy]
python_version = "3.12"
check_untyped_defs = true
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
plugins = [
    "mypy_django_plugin.main",
]

[[tool.mypy.overrides]]
# Django migrations should not produce any errors:
module = "*.migrations.*"
ignore_errors = true

[tool.django-stubs]
django_settings_module = "config.settings.test"

# ==== djLint ====
[tool.djlint]
blank_line_after_tag = "load,extends"
close_void_tags = true
format_css = true
format_js = true
ignore = "H006,H030,H031,T002"
include = "H017,H035"
indent = 2
max_line_length = 119
profile = "django"

[tool.djlint.css]
indent_size = 2

[tool.djlint.js]
indent_size = 2

[tool.ruff]
target-version = "py312"
# Exclude a variety of commonly ignored directories.
extend-exclude = [
    "*/migrations/*.py",
    "staticfiles/*",
    "venv/",
    ".venv/",
    ".cache/",
    "__pycache__/",
    ".pytest_cache/",
    ".mypy_cache/",
    ".ruff_cache/",
    ".git/",
    "node_modules/",
    "build/",
    "dist/",
    "*.egg-info/",
    ".tox/",
    "htmlcov/"
]

[tool.ruff.lint]
select = [
    "F",
    "E",
    "W",
    "C90",
    "I",
    "N",
    "UP",
    "YTT",
    # "ANN", # flake8-annotations: should support this in the future but 100+ errors atm
    "ASYNC",
    "S",
    "BLE",
    "FBT",
    "B",
    "A",
    "COM",
    "C4",
    "DTZ",
    "T10",
    "DJ",
    "EM",
    "EXE",
    "FA",
    'ISC',
    "ICN",
    "G",
    'INP',
    'PIE',
    "T20",
    'PYI',
    'PT',
    "Q",
    "RSE",
    "RET",
    "SLF",
    "SLOT",
    "SIM",
    "TID",
    "TC",
    "INT",
    # "ARG", # Unused function argument
    "PTH",
    "ERA",
    "PD",
    "PGH",
    "PL",
    "TRY",
    "FLY",
    # "NPY",
    # "AIR",
    "PERF",
    # "FURB",
    # "LOG",
    "RUF",
]
ignore = [
    "S101", # Use of assert detected https://docs.astral.sh/ruff/rules/assert/
    "RUF012", # Mutable class attributes should be annotated with `typing.ClassVar`
    "SIM102", # sometimes it's better to nest
    "UP038", # Checks for uses of isinstance/issubclass that take a tuple
    # of types for comparison.
    # Deactivated because it can make the code slow:
    # https://github.com/astral-sh/ruff/issues/7871
    "COM812", # Trailing comma missing - conflicts with ruff formatter
]
# The fixes in extend-unsafe-fixes will require
# provide the `--unsafe-fixes` flag when fixing.
extend-unsafe-fixes = [
  "UP038",
]

[tool.ruff.lint.isort]
force-single-line = true

[dependency-groups]
dev = [
    # Development dependencies
    "Werkzeug[watchdog]==3.1.3",
    "ipdb==0.13.13",
    "psycopg[c]==3.2.9",
    "watchfiles==1.1.1",
    # Testing
    "mypy==1.18.2",
    "django-stubs[compatible-mypy]==5.2.9",
    "pytest==9.0.1",
    "pytest-sugar==1.1.1",
    # Code quality
    "ruff==0.14.14",
    "coverage==7.13.3",
    "djlint==1.36.4",
    "pre-commit==4.5.1",
    # Django development
    "factory-boy==3.3.3",
    "django-debug-toolbar==6.2.0",
    "django-extensions==4.1",
    "django-coverage-plugin==3.2.0",
    "pytest-django==4.11.1",
    "django-browser-reload==1.21.0",
]

production = [
    # Production dependencies
    "gunicorn==25.0.1",
    "psycopg[c]==3.2.9",
    "sentry-sdk==2.51.0",
    # Django production
    "django-anymail[postmark]==14.0",
]

[project]
name = "temp_competibee"
version = "1"
description = "a competibee temp"
readme = "README.md"


requires-python = "==3.12.*"
dependencies = [
    # Base dependencies
    "python-slugify==8.0.4",
    "Pillow==11.3.0",
    "argon2-cffi==25.1.0",
    "whitenoise==6.11.0",
    "redis==7.1.0",
    "hiredis==3.3.0",
    "celery==5.6.0",
    "django-celery-beat==2.8.1",
    "flower==2.0.1",
    # Django
    "django==5.2.5",
    "django-environ==0.12.0",
    "django-model-utils==5.0.0",
    "django-allauth[socialaccount]==65.14.0",
    "django-ninja==1.5.3",
    "django-redis==6.0.0",
    # Stripe
    "stripe==14.3.0",
    # Admin
    "django-unfold==0.73.0",
    # AI
    "posthog==7.8.2",
]
